+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [Unreleased]
+### Added
+- Ask for user confirmation before downloading large files.
+
 ## 0.2.0 - 2025-01-06
 ### Added
 - Add clipping functionality to `get_elev_raster` function to clip the raster to the bounding box.